requirements.txt
# Supply Chain Multi-Agent System Requirements

# Core ML and Data Science
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Data Models
pydantic>=2.5.0

# Optional: For Jupyter notebooks
jupyter>=1.0.0
ipython>=8.18.0

# API (Option 2: Transparent Defaults)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# AWS Bedrock + Strands (semantic classifier + LLM orchestration)
boto3>=1.34.0
strands-agents>=1.7.0

# Development Tools (optional)
black>=23.12.0
pytest>=7.4.0
